requirements.txt
boto3
fastapi
uvicorn
python-dotenv